refactor(docs-framework): replace oclif CLI with library-first runCollect()
Removes the two oclif command wrappers (never invoked — consumers call the
library directly) plus the @oclif/core and @eventuras/lustro-search deps they
pulled in. Adds runCollect(), which ports the command's config discovery,
repo-root detection and validation into a one-line library API, and adopts it
in dev-docs' collect script.
