+++ libs/docs-framework/src/index.ts
@@ -1,3 +1,5 @@
 export { defineDocsConfig } from './collector/config.js';
 export type { DocsConfig, DocSource } from './collector/config.js';
 export { collect } from './collector/collect.js';
+export { runCollect } from './collector/run.js';
+export type { RunCollectOptions } from './collector/run.js';
